+++ actix-web/CHANGES.md
@@ -6,9 +6,11 @@
 - Add `HttpRequest::url_for_map` and `HttpRequest::url_for_iter` methods for named URL parameters. [#3895]
 - Ignore unparsable cookies in `Cookie` request header.
 - Add `experimental-introspection` feature to report configured routes [#3594]
+- Add config/method for `TCP_NODELAY`. [#3918]
 
 [#3895]: https://github.com/actix/actix-web/pull/3895
 [#3594]: https://github.com/actix/actix-web/pull/3594
+[#3918]: https://github.com/actix/actix-web/pull/3918
 
 ## 4.12.1